+++ saqc/funcs/outliers.py
@@ -579,6 +579,9 @@ class OutliersMixin:
         hydrological data. See the notes section for an overview over the algorithms
         basic steps.
 
+            .. deprecated:: 2.6.0
+               Deprecated Function. Please refer to :py:meth:`~saqc.SaQC.flagByStray`.
+
         Parameters
         ----------
         trafo :
@@ -725,9 +728,9 @@ class OutliersMixin:
 
         warnings.warn(
             """
-                FlagMVScores is deprecated and will be removed with Version 2.8.
-                To replicate the function, transform the different fields involved 
-                via explicit applications of some transformations, than calculate the 
+                flagMVScores is deprecated and will be removed with Version 2.8.
+                To replicate the function, transform the different fields involved
+                via explicit applications of some transformations, than calculate the
                 kNN scores via `saqc.SaQC.assignkNScores` and finally assign the STRAY
                 algorithm via `saqc.SaQC.flagByStray`.
                 """,
@@ -861,10 +864,10 @@ class OutliersMixin:
 
         warnings.warn(
             "The function flagRaise is deprecated with no 100% exact replacement function."
-            "When looking for changes in the value course, the use of flagraise can be replicated and more easily aimed "
-            "for, via the method flagJump.\n"
+            "When looking for changes in the value course, the use of flagRaise can be replicated and more "
+            "easily aimed for, via the method flagJump.\n"
             "When looking for raises to outliers or plateaus, use one of: "
-            "flagZScore(outliers), flagUniLOF (outliers and small plateaus) or flagOffset(Plateaus)",
+            "flagZScore (outliers), flagUniLOF (outliers and small plateaus) or flagOffset (plateaus)",
             DeprecationWarning,
         )
 
@@ -970,6 +973,10 @@ class OutliersMixin:
         Flag outiers using the modified Z-score outlier detection method.
 
         See references [1] for more details on the algorithm.
+
+            .. deprecated:: 2.6.0
+               Deprecated Function. Please refer to :py:meth:`~saqc.SaQC.flagZScore`.
+
 
         Note
         ----
@@ -1262,8 +1269,8 @@ class OutliersMixin:
         """
 
         warnings.warn(
-            "The function flagGrubbs is deprecated due to its inferior performance, with no 100% exact replacement function."
-            "When looking for outliers use one of: "
+            "The function flagByGrubbs is deprecated due to its inferior performance, with "
+            "no 100% exact replacement function. When looking for outliers use one of: "
             "flagZScore, flagUniLOF",
             DeprecationWarning,
         )
@@ -1324,85 +1331,6 @@ class OutliersMixin:
 
         self._flags[to_flag, field] = flag
         return self
-
-    @register(
-        mask=["field"],
-        demask=["field"],
-        squeeze=["field"],
-        multivariate=True,
-        handles_target=False,
-        docstring={"field": DOC_TEMPLATES["field"]},
-    )
-    def flagCrossStatistics(
-        self: "SaQC",
-        field: Sequence[str],
-        thresh: float,
-        method: Literal["modZscore", "Zscore"] = "modZscore",
-        flag: float = BAD,
-        **kwargs,
-    ) -> "SaQC":
-        """
-        Function checks for outliers relatively to the "horizontal" input data axis.
-
-        Notes
-        -----
-        The input variables dont necessarily have to be aligned. If the variables are unaligned, scoring
-        and flagging will only be performed on the subset of indices shared among all input variables.
-
-        For :py:attr:`field` :math:`=[f_1,f_2,...,f_N]` and timestamps :math:`[t_1,t_2,...,t_K]`,
-        the following steps are taken for outlier detection:
-
-        1. All timestamps :math:`t_i`, where there is one :math:`f_k`, with :math:`data[f_K]` having no
-           entry at :math:`t_i`, are excluded from the following process (inner join of the :math:`f_i` fields.)
-        2. for every :math:`0 <= i <= K`, the value
-           :math:`m_j = median(\\{data[f_1][t_i], data[f_2][t_i], ..., data[f_N][t_i]\\})` is calculated
-        3. for every :math:`0 <= i <= K`, the set
-           :math:`\\{data[f_1][t_i] - m_j, data[f_2][t_i] - m_j, ..., data[f_N][t_i] - m_j\\}` is tested for
-           outliers with the specified algorithm (:py:attr:`method` parameter).
-
-        Parameters
-        ----------
-        thresh :
-            Threshold which the outlier score of an value must exceed, for being flagged an outlier.
-
-        method :
-            Method used for calculating the outlier scores.
-
-            * ``'modZscore'``: Median based "sigma"-ish approach. See References [1].
-            * ``'Zscore'``: Score values by how many times the standard deviation they differ from the
-              median. See References [1].
-
-
-        References
-        ----------
-        [1] https://www.itl.nist.gov/div898/handbook/eda/section3/eda35h.htm
-        """
-        new_method_string = {
-            "modZscore": "modified",
-            "Zscore": "standard",
-            np.mean: "standard",
-            np.median: "modified",
-        }
-        call = (
-            f"qc.flagZScore(field={field}, window=1, "
-            f"method={new_method_string[method]}, "
-            f"thresh={thresh}, axis=1)"
-        )
-        warnings.warn(
-            f"The method `flagCrossStatistics` is deprecated and will "
-            f"be removed in verion 2.7 of saqc. To achieve the same behavior "
-            f"use:`{call}`",
-            DeprecationWarning,
-        )
-
-        return self.flagZScore(
-            field=field,
-            window=1,
-            method=new_method_string[method],
-            thresh=thresh,
-            axis=1,
-            flag=flag,
-        )
 
     @register(
         mask=["field"],